+++ framework/SceneGraph/Scene/src/ResourceChanges.cpp
@@ -28,13 +28,13 @@ namespace ramses_internal
     template <typename ELEMENTTYPE>
     void putDataArray(SceneActionCollection& action, const std::vector<ELEMENTTYPE>& dataArray)
     {
-        const UInt32 numElements = static_cast<UInt32>(dataArray.size());
+        const uint32_t numElements = static_cast<uint32_t>(dataArray.size());
         action.write(numElements);
         if (numElements > 0u)
         {
             // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast) TODO(tobias) questionable if correct because ELEMENTTYPE not POD
             const Byte* rawData = reinterpret_cast<const Byte*>(dataArray.data());
-            const UInt32 size = numElements * sizeof(ELEMENTTYPE);
+            const uint32_t size = numElements * sizeof(ELEMENTTYPE);
             action.write(rawData, size);
         }
     }
@@ -42,13 +42,13 @@ namespace ramses_internal
     template <typename ELEMENTTYPE>
     void getDataArray(SceneActionCollection::SceneActionReader& action, std::vector<ELEMENTTYPE>& dataArray)
     {
-        UInt32 numElements = 0u;
+        uint32_t numElements = 0u;
         action.read(numElements);
 
         if (numElements > 0u)
         {
             const Byte* rawData = nullptr;
-            UInt32 size = 0u;
+            uint32_t size = 0u;
             action.readWithoutCopy(rawData, size);
 
             dataArray.resize(numElements);
@@ -58,9 +58,9 @@ namespace ramses_internal
     }
 
     template <typename ELEMENTTYPE>
-    UInt estimatePutDataArraySize(const std::vector<ELEMENTTYPE>& dataArray)
+    size_t estimatePutDataArraySize(const std::vector<ELEMENTTYPE>& dataArray)
     {
-        return sizeof(UInt32) + sizeof(ELEMENTTYPE) * dataArray.size();
+        return sizeof(uint32_t) + sizeof(ELEMENTTYPE) * dataArray.size();
     }
 
     void ResourceChanges::putToSceneAction(SceneActionCollection& action) const
@@ -77,7 +77,7 @@ namespace ramses_internal
         getDataArray(action, m_sceneResourceActions);
     }
 
-    UInt ResourceChanges::getPutSizeEstimate() const
+    size_t ResourceChanges::getPutSizeEstimate() const
     {
         return estimatePutDataArraySize(m_resourcesAdded) +
             estimatePutDataArraySize(m_resourcesRemoved) +
